+++ addon/components/object-list-view-row.js
@@ -218,6 +218,15 @@ export default FlexberryBaseComponent.extend({
   */
   recordsLoaded: false,
 
+  /**
+    Level of hierarchy, that already was loaded.
+
+    @property hierarchyLoadedLevel
+    @type Integer
+    @default -1
+  */
+  hierarchyLoadedLevel: -1,
+
   actions: {
     /**
       Show/hide embedded records.
@@ -285,13 +294,13 @@ export default FlexberryBaseComponent.extend({
     if (!this.get('recordsLoaded')) {
       let id = this.get('record.data.id');
       if (id && this.get('inHierarchicalMode')) {
+        let currentLevel = this.get('_currentLevel');
+        let hierarchyLoadedLevel = this.get('hierarchyLoadedLevel');
+        this.sendAction('loadRecords', id, this, 'records', currentLevel > hierarchyLoadedLevel);
         this.set('recordsLoaded', true);
-        if (this.get('_level') > 0) {
-          this.sendAction('loadRecords', id, this, 'records', false);
-        } else {
-          this.sendAction('loadRecords', id, this, 'records', true);
+        if (currentLevel > hierarchyLoadedLevel) {
+          this.set('hierarchyLoadedLevel', currentLevel);
         }
-
       }
     }
   },